update TaskInfo and TaskSchedule data . by laptop 3

+++ app/src/main/java/com/inz/z/note_book/service/ScheduleService.kt
@@ -275,9 +275,10 @@ class ScheduleService : LifecycleService() {
                 intent.action = Constants.AlarmAction.ALARM_BROADCAST_SCHEDULE_ACTION
             }
         }
+        // TODO: 2021/10/18 code maybe have error 
         return PendingIntent.getBroadcast(
             applicationContext,
-            taskInfo.taskRequestCode,
+            taskInfo.taskRequestCode.toInt(),
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT
         )
